feat: 상수 mapper config로 이동
